Capture selector automatically
This updates the capturing behavior to get the full selector rather than just
the components that are part of a Match. It now keeps captures as a list
of dicts with one for each component in the selector. As part of this, the
selector function is now moved to args instead to resolve the circular reference.